Added initial board representation and agent now stores board
The agent has set_current_board() and current_board variabled added to it to stored the current board.
The Matris class has now been set so an initial blank board is printed on game start.
This blank board is stored by the agent in current board.
Subsequent board information is also stored by the agent.