Many updates relating to memory recall, end of game, csv format and error where skyline is overfilled
Update CSV Format.

Random moves now loops so that a rotation with a valid placement is always found.

Previous actions updated in random moves.

Creation of tetromino_to_input() method, which converts the agent tetromino to the appropriate input the the ANN.

Board hard cleared (self.matrix.dict()) at the end of a game.

Error where Matrix is filled more than 2 above the skyline resolved by adding a flag in case of that occurence.

Additional remeberance of moves when tetromino_placement == False